Progress update
In this version, 9 mines were placed in random cells. Cells are
displayed as black squares.
Uncovered squares and mines can't be flagged. 
Under Progress: Finding the right algorithm to uncover multiple cells at
once. (This has been the most time consuming and challenging part of
this project)
Note: Comment lines 104-155 of MyMouseAdapter to be able to click cells
individually.